Move EntityExtension to a decent place
